requirements.txt
cffi==1.14.5
cryptography==3.4.7
future==0.18.2
greenlet==1.1.0
marshmallow==3.12.1
numpy==1.21.0
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
tk==0.1.0
tkintertable==1.3.3
XlsxWriter==1.4.4
